# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PsychGen BR — AI-driven psychometric instrument development
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/pipeline:
    get:
      operationId: getProjectPipeline
      tags: [projects]
      summary: Pipeline overview for a project (all stages and jobs)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Pipeline overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineOverview"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/items:
    get:
      operationId: listProjectItems
      tags: [items]
      summary: List items for a project
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Project:
      type: object
      required:
        - id
        - name
        - construct
        - language
        - targetAudience
        - status
        - itemCount
        - approvedCount
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        name: { type: string }
        construct: { type: string }
        description:
          type: ["string", "null"]
        language: { type: string }
        targetAudience: { type: string }
        publisher:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, generating, calibrating, ready, archived]
        itemCount: { type: integer }
        approvedCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateProjectInput:
      type: object
      required: [name, construct, targetAudience]
      properties:
        name: { type: string, minLength: 1 }
        construct: { type: string, minLength: 1 }
        description:
          type: ["string", "null"]
        language: { type: string, default: pt-BR }
        targetAudience: { type: string, minLength: 1 }
        publisher:
          type: ["string", "null"]
    UpdateProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        construct: { type: string, minLength: 1 }
        description:
          type: ["string", "null"]
        language: { type: string }
        targetAudience: { type: string }
        publisher:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, generating, calibrating, ready, archived]
...
    PipelineJob:
      type: object
      required:
        - id
        - projectId
        - stage
        - status
        - progress
        - paramsJson
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        projectId: { type: integer }
        stage:
          type: string
          enum: [aigenie, difficulty, irt, sample_design]
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        progress: { type: number, minimum: 0, maximum: 1 }
        message:
          type: ["string", "null"]
        paramsJson:
          type: object
          additionalProperties: true
        resultJson:
          type: ["object", "null"]
          additionalProperties: true
        error:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        finishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PipelineOverview:
      type: object
      required: [projectId, stages]
      properties:
        projectId: { type: integer }
        stages:
          type: array
          items:
            type: object
            required: [stage, status, jobCount]
            properties:
              stage:
                type: string
                enum: [aigenie, difficulty, irt, sample_design]
              status:
                type: string
                enum: [not_started, running, completed, failed]
              jobCount: { type: integer }
              latestJob:
                anyOf:
                  - { $ref: "#/components/schemas/PipelineJob" }
                  - { type: "null" }
    Item:
      type: object
      required:
        - id
        - projectId
        - text
        - status
        - generatedBy
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        projectId: { type: integer }
        text: { type: string }
        construct:
          type: ["string", "null"]
        dimension:
          type: ["string", "null"]
        status:
          type: string
          enum: [generated, needs_review, approved, rejected]
        generatedBy: { type: string, description: "Model that generated the item" }
        difficultyPredicted:
          type: ["number", "null"]
        difficultyEstimated:
          type: ["number", "null"]
        discrimination:
          type: ["number", "null"]
        guessing:
          type: ["number", "null"]
        egaCommunity:
          type: ["integer", "null"]
        humanNotes:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Report:
      type: object
      required: [id, projectId, kind, summary, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        kind:
          type: string
          enum: [aigenie, difficulty, irt, validation, sample_design]
        summary: { type: string }
        metricsJson:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }