# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PsychGen BR — AI-driven psychometric instrument development
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  construct: string;
  /** @nullable */
  description?: string | null;
  language?: string;
  /** @minLength 1 */
  targetAudience: string;
  /** @nullable */
  publisher?: string | null;
}